requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pdfplumber
google-genai
python-dotenv
